Backend/requirements.txt
Django>=5.1.6
django-cors-headers>=4.7.0
djangorestframework>=3.16.0
python-dotenv>=1.0.0 